Turbo midi fix

+++ avr/cores/megacommand/Midi/MidiUartParent.hh
@@ -30,7 +30,7 @@ public:
   uint8_t running_status;
   uint8_t currentChannel;
   uint8_t uart_port;
-  uint8_t speed;
+  uint32_t speed;
 
   bool useRunningStatus;
 	uint16_t sendActiveSenseTimer;
